refactor: delete the host's own code editor, which the editor-tab plugin replaced
The host carried a second, complete code editor alongside the one users
actually get. `CodeEditorUI` with its own Kotlin keyword list, syntax
highlighter, gutter, font and colour hooks and tab component -- 786 lines
across three files -- sitting next to editor-tab, which bundles BossEditor and
owns every editing surface in the product.

It was already understood to be dead weight. Its registration was commented out
in two places in `DefaultPlugin.kt`, and both remaining files said so in their
own comments: "the editor-tab plugin provides the real editor",
"CodeEditorContent only backs the basic shared fallback editor; the editor-tab
plugin renders its own BossEditor and doesn't call it."

Worse than unused, the one path that could still reach it was actively wrong.
`EditorContentProviderImpl.CodeEditorContent` accepted `onSaveRequested`,
`onModifiedStateChange` and `onRunFunction` and discarded all three, so any
caller would have got an editor that silently refused to save. Its own KDoc
recorded this: "effectively VIEW-ONLY".

## What was verified before deleting

- Nothing in BossConsole or `plugin-platform` calls `CodeEditorContent`. The
  only two occurrences are the override and its own doc comment.
- Nothing in the `boss_plugins` workspace calls it either -- across all 30+
  plugin repos the single hit is the interface declaration itself.
- `PlatformCodeEditorUI`, `CodeEditorTabComponent`, `SyntaxHighlightTransformation`,
  the eight `getCodeEditor*` hooks and the single-argument `readFileContent`
  have no reference outside the deleted pair.

## What stays, and why

`EditorContentProviderImpl` itself is live and keeps everything editor-tab
reaches through it: `EditorTabMcpTools` calls `readFileContent`,
`writeFileContent` and `detectLanguage`, which are the `editor_read_file`,
`editor_write_file` and `editor_detect_language` MCP tools.

So the file I/O those rest on had to survive the file it happened to live in.
It moved to `providers/EditorFileIo.kt` next to its only caller, unchanged.
`EditorLanguages` stays where it is, still backing `detectLanguage`, still
guarded against drift with the plugin's own copy by EditorLanguageDetectionTest.

`CodeEditorContent` cannot be dropped -- it is abstract on
`EditorContentProvider`, which lives in the `boss-plugin-api` repo, and removing
it would be a breaking API change across that chain. It now renders one line
naming the plugin that owns the editor, which is strictly better than an editor
that pretends to save.

The `FileReadResult`/`InternalFileReadResult` import alias is gone with it: the
internal type is `FileReadOutcome` now, so it no longer collides with the
plugin-api type of the same name.

## Detekt

15 baseline entries died with the deleted files; 1 was added for the relocated
I/O, whose two generic catches now share a file. Net 14 fewer suppressions.

786 lines of editor deleted, 118 kept. Compiles, `desktopTest` green, detekt
and ktlint green.
